package structure refactored for medialibrary app is working
