+++ app/src/main/java/com/practicum/playlistmaker/medialibrary/di/ViewModelModule.kt
@@ -1,8 +1,8 @@
 package com.practicum.playlistmaker.medialibrary.di
 
 import com.google.gson.Gson
-import com.practicum.playlistmaker.medialibrary.presentation.viewmodel.FavoriteTracksViewModel
-import com.practicum.playlistmaker.medialibrary.presentation.viewmodel.MediaLibraryViewModel
+import com.practicum.playlistmaker.medialibrary.presentation.favorite_tracks.viewmodel.FavoriteTracksViewModel
+import com.practicum.playlistmaker.medialibrary.presentation.root.viewmodel.MediaLibraryViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
