+++ app/src/main/java/com/practicum/playlistmaker/medialibrary/presentation/playlists/viewmodel/PlayListsViewModel.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.medialibrary.presentation.playlists.viewmodel
+
+
+import androidx.lifecycle.ViewModel
+
+
+class PlayListsViewModel(): ViewModel()
