--- app/src/main/java/com/practicum/playlistmaker/medialibrary/presentation/viewmodel/PlayListsViewModel.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.medialibrary.presentation.viewmodel
-
-
-import androidx.lifecycle.ViewModel
-
-
-class PlayListsViewModel(): ViewModel()
